requirements.txt
discord.py[voice]
youtube_dl
beautifulsoup4
google